Seperate xcm config&Refactor accordingly
