+++ runtime/vanilla/src/lib.rs
@@ -21,18 +21,17 @@
 include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 
 mod weights;
+pub mod xcm_config;
 
 use codec::{Decode, Encode, MaxEncodedLen};
 use scale_info::TypeInfo;
 
 use frame_support::{
     dispatch::Weight,
-    log, match_type,
+    log,
     traits::{
-        fungibles::{InspectMetadata, Mutate},
-        tokens::BalanceConversion,
-        ChangeMembers, Contains, EnsureOneOf, EqualPrivilegeOnly, Everything, InstanceFilter,
-        Nothing, OnRuntimeUpgrade,
+        fungibles::InspectMetadata, tokens::BalanceConversion, ChangeMembers, Contains,
+        EnsureOneOf, EqualPrivilegeOnly, InstanceFilter, OnRuntimeUpgrade,
     },
     PalletId,
 };
@@ -41,14 +40,11 @@ use frame_system::{
     EnsureRoot, EnsureSigned,
 };
 
-use orml_traits::{parameter_type_with_key, DataProvider, DataProviderExtended};
-use orml_xcm_support::{IsNativeConcrete, MultiNativeAsset};
-use polkadot_parachain::primitives::Sibling;
+use orml_traits::{DataProvider, DataProviderExtended};
 use polkadot_runtime_common::SlowAdjustingFeeUpdate;
 use primitives::{
-    currency::MultiCurrencyAdapter,
     network::HEIKO_PREFIX,
-    tokens::{EUSDC, EUSDT, HKO, KAR, KSM, KUSD, LKSM, SKSM},
+    tokens::{EUSDC, EUSDT, HKO, KSM, SKSM},
     Index, *,
 };
 use sp_api::impl_runtime_apis;
@@ -60,7 +56,7 @@ use sp_runtime::{
     create_runtime_str, generic, impl_opaque_keys,
     traits::{
         self, AccountIdConversion, AccountIdLookup, BlakeTwo256, Block as BlockT,
-        BlockNumberProvider, Convert, Zero,
+        BlockNumberProvider, Zero,
     },
     transaction_validity::{TransactionSource, TransactionValidity},
     ApplyExtrinsicResult, DispatchError, FixedU128, KeyTypeId, Perbill, Permill, RuntimeDebug,
@@ -70,15 +66,6 @@ use sp_std::prelude::*;
 #[cfg(feature = "std")]
 use sp_version::NativeVersion;
 use sp_version::RuntimeVersion;
-use xcm::latest::prelude::*;
-use xcm_builder::{
-    AccountId32Aliases, AllowKnownQueryResponses, AllowSubscriptionsFrom,
-    AllowTopLevelPaidExecutionFrom, EnsureXcmOrigin, FixedRateOfFungible, FixedWeightBounds,
-    LocationInverter, ParentAsSuperuser, ParentIsPreset, RelayChainAsNative,
-    SiblingParachainAsNative, SiblingParachainConvertsVia, SignedAccountId32AsNative,
-    SignedToAccountId32, SovereignSignedViaLocation, TakeRevenue, TakeWeightCredit,
-};
-use xcm_executor::{Config, XcmExecutor};
 
 pub mod constants;
 pub mod impls;
@@ -107,7 +94,6 @@ pub use frame_support::{
     },
     StorageValue,
 };
-use pallet_xcm::XcmPassthrough;
 use time::*;
 
 /// Opaque types. These are used by the CLI to instantiate machinery that don't need to know
@@ -333,155 +319,8 @@ parameter_types! {
    pub TreasuryAccount: AccountId = TreasuryPalletId::get().into_account();
 }
 
-impl orml_xcm::Config for Runtime {
-    type Event = Event;
-    type SovereignOrigin = EnsureRootOrMoreThanHalfGeneralCouncil;
-}
-
 parameter_types! {
     pub const LoansPalletId: PalletId = PalletId(*b"par/loan");
-}
-
-pub struct CurrencyIdConvert;
-impl Convert<CurrencyId, Option<MultiLocation>> for CurrencyIdConvert {
-    fn convert(id: CurrencyId) -> Option<MultiLocation> {
-        match id {
-            KSM => Some(MultiLocation::parent()),
-            SKSM => Some(MultiLocation::new(
-                1,
-                X2(
-                    Parachain(ParachainInfo::parachain_id().into()),
-                    GeneralKey(b"sKSM".to_vec()),
-                ),
-            )),
-            HKO => Some(MultiLocation::new(
-                1,
-                X2(
-                    Parachain(ParachainInfo::parachain_id().into()),
-                    GeneralKey(b"HKO".to_vec()),
-                ),
-            )),
-            KAR => Some(MultiLocation::new(
-                1,
-                X2(
-                    Parachain(paras::karura::ID),
-                    GeneralKey(paras::karura::KAR_KEY.to_vec()),
-                ),
-            )),
-            KUSD => Some(MultiLocation::new(
-                1,
-                X2(
-                    Parachain(paras::karura::ID),
-                    GeneralKey(paras::karura::KUSD_KEY.to_vec()),
-                ),
-            )),
-            LKSM => Some(MultiLocation::new(
-                1,
-                X2(
-                    Parachain(paras::karura::ID),
-                    GeneralKey(paras::karura::LKSM_KEY.to_vec()),
-                ),
-            )),
-            _ => None,
-        }
-    }
-}
-
-impl Convert<MultiLocation, Option<CurrencyId>> for CurrencyIdConvert {
-    fn convert(location: MultiLocation) -> Option<CurrencyId> {
-        match location {
-            MultiLocation {
-                parents: 1,
-                interior: Here,
-            } => Some(KSM),
-            MultiLocation {
-                parents: 1,
-                interior: X2(Parachain(id), GeneralKey(key)),
-            } if ParaId::from(id) == ParachainInfo::parachain_id() && key == b"sKSM".to_vec() => {
-                Some(SKSM)
-            }
-            MultiLocation {
-                parents: 0,
-                interior: X1(GeneralKey(key)),
-            } if key == b"sKSM".to_vec() => Some(SKSM),
-            MultiLocation {
-                parents: 1,
-                interior: X2(Parachain(id), GeneralKey(key)),
-            } if ParaId::from(id) == ParachainInfo::parachain_id() && key == b"HKO".to_vec() => {
-                Some(HKO)
-            }
-            MultiLocation {
-                parents: 0,
-                interior: X1(GeneralKey(key)),
-            } if key == b"HKO".to_vec() => Some(HKO),
-            MultiLocation {
-                parents: 1,
-                interior: X2(Parachain(id), GeneralKey(key)),
-            } if id == paras::karura::ID && key == paras::karura::KUSD_KEY.to_vec() => Some(KUSD),
-            MultiLocation {
-                parents: 1,
-                interior: X2(Parachain(id), GeneralKey(key)),
-            } if id == paras::karura::ID && key == paras::karura::KAR_KEY.to_vec() => Some(KAR),
-            MultiLocation {
-                parents: 1,
-                interior: X2(Parachain(id), GeneralKey(key)),
-            } if id == paras::karura::ID && key == paras::karura::LKSM_KEY.to_vec() => Some(LKSM),
-            _ => None,
-        }
-    }
-}
-
-impl Convert<MultiAsset, Option<CurrencyId>> for CurrencyIdConvert {
-    fn convert(a: MultiAsset) -> Option<CurrencyId> {
-        if let MultiAsset {
-            id: AssetId::Concrete(id),
-            fun: _,
-        } = a
-        {
-            Self::convert(id)
-        } else {
-            None
-        }
-    }
-}
-
-pub struct AccountIdToMultiLocation;
-impl Convert<AccountId, MultiLocation> for AccountIdToMultiLocation {
-    fn convert(account_id: AccountId) -> MultiLocation {
-        X1(AccountId32 {
-            network: NetworkId::Any,
-            id: account_id.into(),
-        })
-        .into()
-    }
-}
-
-parameter_types! {
-    pub SelfLocation: MultiLocation = MultiLocation::new(1, X1(Parachain(ParachainInfo::parachain_id().into())));
-    pub const BaseXcmWeight: Weight = 150_000_000;
-    pub const MaxInstructions: u32 = 100;
-    pub const MaxAssetsForTransfer: usize = 2;
-}
-
-parameter_type_with_key! {
-    pub ParachainMinFee: |_location: MultiLocation| -> u128 {
-        u128::MAX
-    };
-}
-
-impl orml_xtokens::Config for Runtime {
-    type Event = Event;
-    type Balance = Balance;
-    type CurrencyId = CurrencyId;
-    type CurrencyIdConvert = CurrencyIdConvert;
-    type AccountIdToMultiLocation = AccountIdToMultiLocation;
-    type SelfLocation = SelfLocation;
-    type XcmExecutor = XcmExecutor<XcmConfig>;
-    type Weigher = FixedWeightBounds<BaseXcmWeight, Call, MaxInstructions>;
-    type BaseXcmWeight = BaseXcmWeight;
-    type LocationInverter = LocationInverter<Ancestry>;
-    type MaxAssetsForTransfer = MaxAssetsForTransfer;
-    type MinXcmFee = ParachainMinFee;
 }
 
 parameter_types! {
@@ -840,98 +679,6 @@ impl pallet_utility::Config for Runtime {
     type WeightInfo = pallet_utility::weights::SubstrateWeight<Runtime>;
 }
 
-/// Local origins on this chain are allowed to dispatch XCM sends/executions. However, we later
-/// block this via `ExecuteXcmOrigin`.
-pub type LocalOriginToLocation = SignedToAccountId32<Origin, AccountId, RelayNetwork>;
-
-/// The means for routing XCM messages which are not for local execution into the right message
-/// queues.
-pub type XcmRouter = (
-    // Two routers - use UMP to communicate with the relay chain:
-    cumulus_primitives_utility::ParentAsUmp<ParachainSystem, PolkadotXcm>,
-    // ..and XCMP to communicate with the sibling chains.
-    XcmpQueue,
-);
-
-impl pallet_xcm::Config for Runtime {
-    const VERSION_DISCOVERY_QUEUE_SIZE: u32 = 100;
-
-    type Origin = Origin;
-    type Call = Call;
-    type Event = Event;
-    type SendXcmOrigin = EnsureXcmOrigin<Origin, LocalOriginToLocation>;
-    type XcmRouter = XcmRouter;
-    type ExecuteXcmOrigin = EnsureXcmOrigin<Origin, LocalOriginToLocation>;
-    type XcmExecuteFilter = Nothing;
-    type XcmReserveTransferFilter = Everything;
-    type XcmExecutor = XcmExecutor<XcmConfig>;
-    // Teleporting is disabled.
-    type XcmTeleportFilter = Nothing;
-    type Weigher = FixedWeightBounds<BaseXcmWeight, Call, MaxInstructions>;
-    type LocationInverter = LocationInverter<Ancestry>;
-    type AdvertisedXcmVersion = pallet_xcm::CurrentXcmVersion;
-}
-
-impl cumulus_pallet_xcm::Config for Runtime {
-    type Event = Event;
-    type XcmExecutor = XcmExecutor<XcmConfig>;
-}
-
-impl cumulus_pallet_xcmp_queue::Config for Runtime {
-    type Event = Event;
-    type XcmExecutor = XcmExecutor<XcmConfig>;
-    type ExecuteOverweightOrigin = EnsureRootOrMoreThanHalfGeneralCouncil;
-    type ChannelInfo = ParachainSystem;
-    type VersionWrapper = PolkadotXcm;
-    type ControllerOrigin = EnsureRootOrMoreThanHalfGeneralCouncil;
-    type ControllerOriginConverter = XcmOriginToTransactDispatchOrigin;
-}
-
-impl cumulus_pallet_dmp_queue::Config for Runtime {
-    type Event = Event;
-    type XcmExecutor = XcmExecutor<XcmConfig>;
-    type ExecuteOverweightOrigin = EnsureRootOrMoreThanHalfGeneralCouncil;
-}
-
-parameter_types! {
-    pub const ReservedXcmpWeight: Weight = MAXIMUM_BLOCK_WEIGHT / 4;
-    pub const ReservedDmpWeight: Weight =  MAXIMUM_BLOCK_WEIGHT / 4;
-}
-
-impl cumulus_pallet_parachain_system::Config for Runtime {
-    type Event = Event;
-    type OnSystemEvent = ();
-    type SelfParaId = ParachainInfo;
-    type DmpMessageHandler = DmpQueue;
-    type OutboundXcmpMessageSource = XcmpQueue;
-    type XcmpMessageHandler = XcmpQueue;
-    type ReservedXcmpWeight = ReservedXcmpWeight;
-    type ReservedDmpWeight = ReservedDmpWeight;
-}
-
-impl parachain_info::Config for Runtime {}
-
-parameter_types! {
-    pub const RelayLocation: MultiLocation = MultiLocation::parent();
-    pub RelayNetwork: NetworkId = NetworkId::Kusama;
-    pub RelayCurrency: CurrencyId = KSM;
-    pub VanillaNetwork: NetworkId = NetworkId::Named("vanilla".into());
-    pub RelayChainOrigin: Origin = cumulus_pallet_xcm::Origin::Relay.into();
-    pub Ancestry: MultiLocation = MultiLocation::new(0, X1(Parachain(ParachainInfo::parachain_id().into())));
-}
-
-/// Type for specifying how a `MultiLocation` can be converted into an `AccountId`. This is used
-/// when determining ownership of accounts for asset transacting and when attempting to use XCM
-/// `Transact` in order to determine the dispatch Origin.
-pub type LocationToAccountId = (
-    // The parent (Relay-chain) origin converts to the default `AccountId`.
-    ParentIsPreset<AccountId>,
-    // Sibling parachain origins convert to AccountId via the `ParaId::into`.
-    SiblingParachainConvertsVia<Sibling, AccountId>,
-    // Straight up local `AccountId32` origins just alias directly to `AccountId`.
-    AccountId32Aliases<RelayNetwork, AccountId>,
-);
-
 parameter_types! {
     pub const NativeCurrencyId: CurrencyId = NATIVE_ASSET_ID;
     pub GiftAccount: AccountId = PalletId(*b"par/gift").into_account();
@@ -953,161 +700,6 @@ impl BalanceConversion<Balance, CurrencyId, Balance> for GiftConvert {
             _ => Zero::zero(),
         })
     }
-}
-
-/// Means for transacting assets on this chain.
-pub type LocalAssetTransactor = MultiCurrencyAdapter<
-    // Use this currency:
-    CurrencyAdapter,
-    // Use this currency when it is a fungible asset matching the given location or name:
-    IsNativeConcrete<CurrencyId, CurrencyIdConvert>,
-    // Our chain's account ID type (we can't get away without mentioning it explicitly):
-    AccountId,
-    Balance,
-    // Do a simple punn to convert an AccountId32 MultiLocation into a native chain account ID:
-    LocationToAccountId,
-    CurrencyIdConvert,
-    NativeCurrencyId,
-    ExistentialDeposit,
-    GiftAccount,
-    GiftConvert,
->;
-
-/// This is the type we use to convert an (incoming) XCM origin into a local `Origin` instance,
-/// ready for dispatching a transaction with Xcm's `Transact`. There is an `OriginKind` which can
-/// biases the kind of local `Origin` it will become.
-pub type XcmOriginToTransactDispatchOrigin = (
-    // Sovereign account converter; this attempts to derive an `AccountId` from the origin location
-    // using `LocationToAccountId` and then turn that into the usual `Signed` origin. Useful for
-    // foreign chains who want to have a local sovereign account on this chain which they control.
-    SovereignSignedViaLocation<LocationToAccountId, Origin>,
-    // Native converter for Relay-chain (Parent) location; will converts to a `Relay` origin when
-    // recognised.
-    RelayChainAsNative<RelayChainOrigin, Origin>,
-    // Native converter for sibling Parachains; will convert to a `SiblingPara` origin when
-    // recognised.
-    SiblingParachainAsNative<cumulus_pallet_xcm::Origin, Origin>,
-    // Superuser converter for the Relay-chain (Parent) location. This will allow it to issue a
-    // transaction from the Root origin.
-    ParentAsSuperuser<Origin>,
-    // Native signed account converter; this just converts an `AccountId32` origin into a normal
-    // `Origin::Signed` origin of the same 32-byte value.
-    SignedAccountId32AsNative<RelayNetwork, Origin>,
-    // Xcm origins can be represented natively under the Xcm pallet's Xcm origin.
-    XcmPassthrough<Origin>,
-);
-
-parameter_types! {
-    pub KsmPerSecond: (AssetId, u128) = (AssetId::Concrete(MultiLocation::parent()), ksm_per_second());
-    pub SKSMPerSecond: (AssetId, u128) = (
-        MultiLocation::new(
-            1,
-            X2(Parachain(ParachainInfo::parachain_id().into()), GeneralKey(b"sKSM".to_vec())),
-        ).into(),
-        ksm_per_second()
-    );
-    pub SKSMPerSecondOfCanonicalLocation: (AssetId, u128) = (
-        MultiLocation::new(
-            0,
-            X1(GeneralKey(b"sKSM".to_vec())),
-        ).into(),
-        ksm_per_second()
-    );
-    pub HkoPerSecond: (AssetId, u128) = (
-        MultiLocation::new(
-            1,
-            X2(Parachain(ParachainInfo::parachain_id().into()), GeneralKey(b"HKO".to_vec())),
-        ).into(),
-        ksm_per_second() * 30
-    );
-    pub HkoPerSecondOfCanonicalLocation: (AssetId, u128) = (
-        MultiLocation::new(
-            0,
-            X1(GeneralKey(b"HKO".to_vec())),
-        ).into(),
-        ksm_per_second() * 30
-    );
-    pub KusdPerSecond: (AssetId, u128) = (
-        MultiLocation::new(
-            1,
-            X2(Parachain(paras::karura::ID), GeneralKey(paras::karura::KUSD_KEY.to_vec())),
-        ).into(),
-        ksm_per_second() * 400
-    );
-    pub KarPerSecond: (AssetId, u128) = (
-        MultiLocation::new(
-            1,
-            X2(Parachain(paras::karura::ID), GeneralKey(paras::karura::KAR_KEY.to_vec())),
-        ).into(),
-        ksm_per_second() * 50
-    );
-    pub LKSMPerSecond: (AssetId, u128) = (
-        MultiLocation::new(
-            1,
-            X2(Parachain(paras::karura::ID), GeneralKey(paras::karura::LKSM_KEY.to_vec())),
-        ).into(),
-        ksm_per_second()
-    );
-}
-
-match_type! {
-    pub type ParentOrSiblings: impl Contains<MultiLocation> = {
-        MultiLocation { parents: 1, interior: Here } |
-        MultiLocation { parents: 1, interior: X1(_) }
-    };
-}
-
-pub type Barrier = (
-    TakeWeightCredit,
-    AllowKnownQueryResponses<PolkadotXcm>,
-    AllowSubscriptionsFrom<ParentOrSiblings>,
-    AllowTopLevelPaidExecutionFrom<Everything>,
-);
-
-pub struct ToTreasury;
-impl TakeRevenue for ToTreasury {
-    fn take_revenue(revenue: MultiAsset) {
-        if let MultiAsset {
-            id: AssetId::Concrete(id),
-            fun: Fungibility::Fungible(amount),
-        } = revenue
-        {
-            if let Some(currency_id) = CurrencyIdConvert::convert(id) {
-                let _ = Assets::mint_into(currency_id, &TreasuryAccount::get(), amount);
-            }
-        }
-    }
-}
-
-pub type Trader = (
-    FixedRateOfFungible<KsmPerSecond, ToTreasury>,
-    FixedRateOfFungible<SKSMPerSecond, ToTreasury>,
-    FixedRateOfFungible<SKSMPerSecondOfCanonicalLocation, ToTreasury>,
-    FixedRateOfFungible<HkoPerSecond, ToTreasury>,
-    FixedRateOfFungible<HkoPerSecondOfCanonicalLocation, ToTreasury>,
-    FixedRateOfFungible<KusdPerSecond, ToTreasury>,
-    FixedRateOfFungible<KarPerSecond, ToTreasury>,
-    FixedRateOfFungible<LKSMPerSecond, ToTreasury>,
-);
-
-pub struct XcmConfig;
-impl Config for XcmConfig {
-    type Call = Call;
-    type XcmSender = XcmRouter;
-    // How to withdraw and deposit an asset.
-    type AssetTransactor = LocalAssetTransactor;
-    type OriginConverter = XcmOriginToTransactDispatchOrigin;
-    type IsReserve = MultiNativeAsset;
-    // Teleporting is disabled.
-    type IsTeleporter = ();
-    type LocationInverter = LocationInverter<Ancestry>;
-    type Barrier = Barrier;
-    type Weigher = FixedWeightBounds<BaseXcmWeight, Call, MaxInstructions>;
-    type Trader = Trader;
-    type ResponseHandler = PolkadotXcm;
-    type SubscriptionService = PolkadotXcm;
-    type AssetTrap = PolkadotXcm;
-    type AssetClaims = PolkadotXcm;
 }
 
 parameter_types! {
@@ -1563,6 +1155,10 @@ impl<T: cumulus_pallet_parachain_system::Config> BlockNumberProvider
     }
 }
 
+parameter_types! {
+    pub RelayCurrency: CurrencyId = KSM;
+}
+
 impl pallet_crowdloans::Config for Runtime {
     type Event = Event;
     type Origin = Origin;
@@ -1587,25 +1183,6 @@ impl pallet_crowdloans::Config for Runtime {
     type XCM = XcmHelper;
     type RelayChainBlockNumberProvider = RelayChainBlockNumberProvider<Runtime>;
     type Members = CrowdloansAutomatorsMembership;
-}
-
-parameter_types! {
-    pub const XcmHelperPalletId: PalletId = PalletId(*b"par/fees");
-    pub const NotifyTimeout: BlockNumber = 100;
-}
-
-impl pallet_xcm_helper::Config for Runtime {
-    type Event = Event;
-    type UpdateOrigin = EnsureRootOrMoreThanHalfGeneralCouncil;
-    type Assets = Assets;
-    type XcmSender = XcmRouter;
-    type RelayNetwork = RelayNetwork;
-    type PalletId = XcmHelperPalletId;
-    type NotifyTimeout = NotifyTimeout;
-    type AccountIdToMultiLocation = AccountIdToMultiLocation;
-    type RefundLocation = RefundLocation;
-    type BlockNumberProvider = frame_system::Pallet<Runtime>;
-    type WeightInfo = pallet_xcm_helper::weights::SubstrateWeight<Runtime>;
 }
 
 parameter_types! {
